Allow for saving and loading shitposting databases
In this way it's not necessary to rebuild the database on each run
(however it must be trained manually after the first training).

The `board` variable now contains the board name without the `/`s, to be
able to name the database files.

`{board}-data` and `{board}-images` files are created for each trained
board.